backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
python-decouple==3.8
Pillow==11.1.0
openpyxl==3.1.5
reportlab==4.2.5
